changed how end of line intersections are handled.
if the other line intersects the end of self then it is no longer counted
as an intersection. This fixes the problem where if a line went through
the start of one line and the end of another line (a line passes through
a node of the shape) it used to be counted as two intersections but should
really only be one.