segmentsIntersect() finally fixed, int division and end check
denom was assumed to be an int if the inputs were int which was causing
an integer division error for t and u.

the test for end point intersection has been fixed so that if the "end"
(as apposed to self) is intersected then this is not counted has an
intersection. both ends of "other" and counted as hits.