Shape now is oriented properly and isInside is fixed
The shape is now oriented so that the left of a line is inside the
shape. This makes most operations easier. For instance now getOffsetLine
give you the line you want right away instead of sending back both and
making Shape decide. Some additional changes were made to isInside() so
that 'tangent' points are now handled properly. Before if our test line
went through the end points of the shape we might get the wrong results.
Now that is fixed by throwing away the bad test line and drawing a new
one in a different direction and testing again.