layerParam now used as key for layers
I found a typo when adding a new parameter to the layer key. Since only
layerHeight in layerParam does not make a layer unique I decided it was
better to have the possibility of recalculating a layer with all
parameters the same except layer height than it was to accidently miss a
parameter which might make a layer unique.