just going to try to make the 3D model in a separate program
Spent most of the day trying to get gui.py to work with showing the
matplotlib graph. Tried many things, lots of googling, might have broken
python on the computer, but nothing seemed to work. I've resigned to
trying to set the program up to make the 3D model in a separate program
and integrate it into gui.py at a later date.