improved model, changing string to variable
Changed which function in figura creates the text file of data points,
now has multiple layers with different z-values. Improved the scales in
the 3D model to show tick intervals. About to change the string names of
variables in gui.py to be actual variables. Made saving in gui.py more
pythonic.