changed what saves to text file, changing when data is read in
Added lines to text file that tell when a new layer is being made and
what layer it is. Realized that the Page_Model page is being initialized
right away, thus using the old version of the data_points.txt file, so
I'm beginning to change when it reads the data.