Extrusion Rate is now calculated with nozzleDiameter
If you wanted to make a part with a low infill percentage, had a large
distance between paths, the previous calculation would have tried to
pump out a lot of material. Instead we now use the nozzle diameter to
calculate the extrusion rate which allows us to independently control
the path width and the extrusion rate.