actual "made saving more understandable, fixed 3D Model page"
Messed up the original commit, didn't include the program that was
changed. Again, changed some variable names in the save method to make
it more understandable. Can now type the .json or .gcode when saving a
file (only works if using one of those two extensions). Fixed a bug in
the 3D Model page that was due to a naming error.