created two versions of gui
gui.3D uses pygame and openGL to create the 3D model. this is more
detailed and faster. However, you must import pygame and openGL for it
to work, thus you must have pygame and openGL downloaded for it to work.

gui.py uses matplotlib for the 3D model. Not the best, but it gets the
job done. save, upload, generate code, and model making functionality
has been tested and it should all work with the current state of the
program (had to fix some stuff now that stl_path is a thing)