Include return type for all methods and add straightLines()
Infill uses a None input as a default to then call lineField() and create a field of parallel lines. It uses extra information about the trimShape to efficiently create an appropriately sized field. The straightLines method is added here to provide the option in the gui dropdown. I have marked its return type as LineGroup to help show that this is an Infill and not an Outline. Since it returns None this is probably an incorrect way to handle the problem.