pattern variable info now shows up
Got the program to perform a second shift to allow space for info of
pattern variables. Correctly displays labels, created an additional
label in the 0th column to display if the data is outline or pattern.
Currently has bug where it deletes pattern labels instead of outline
labels when switching pattern from an inline to 'none'.