created dropdowns namedtuple, fixed doneshapes typo
Created Dropdown namedtuple for dropdowns. Moved outline and pattern
into dropdowns. In the middle of replacing all of the ('outline',
'pattern') iterations with iterating through self.dropdowns. Got rid of
two constants that are no longer needed. Fixed a typo in doneshapes that
caused a method to be shifted over.